Basic splitting
