+++ src/geotiff-validator/Dockerfile
@@ -11,6 +11,7 @@ RUN chmod +x /home/start-worker.sh /home/wait-for.sh
 WORKDIR /home/
 RUN npm install
 COPY src/geotiff-validator/index.js worker/
+COPY src/geotiff-validator/validator.js worker/
 COPY src/workerTemplate.js .
 ADD src/geotiff-validator/config config
